Update github-pages from 207 to 214 (#42)
This addresses a security vulnerability in kramdown.
I also updated Ruby from 2.6.5 to 2.7.2, and Bundler to 2.2.13. There’s no reason to stay on old versions. Ruby 3.0.0 is out, but Jekyll is not yet fully compatible with it.

Eventually, I’d like us to get rid of the `github-pages` gem. It’s not needed to deploy to GitHub Pages, and it’s preventing us from using the latest version of Jekyll.

Here’s a step-by-step guide I wrote for deploying to GitHub Pages using GitHub Actions:

https://www.moncefbelyamani.com/making-github-pages-work-with-latest-jekyll/